Local and MythTV bookmark support
A preference setting allows selection of local, mythtv, or auto selection.
Local will store the bookmarks on a local database. This works always.
MythTV will store the bookmark as a MythTV bookmark on the server. This
only works where there is a seektable present.
Auto selection will store bookmark on MythTV and if it cannot, it will use the
local database.