Add LiveTV support and Stop Recording menu item.
Display TV Channels in the LiveTV group.
Allow selecting and playing a channel.
To play Live TV, the system schedules a recoding on teh channel and then plays it.
Exiting from playback stops the recording and deletes the recording rule.
Also add "Stop Recording" to the menu whenever a recording in progress is viewed.