Fix for unplayable recordings from the UK
The recordings that would not play have an empty audio track.
This was causing exoplayer to die because it would not mark the
stream as having been "prepared".

The change will mark this type of track as prepared audio so that
playback can continue.

It only allows continuation if at least one valid video and one
valid audio track were found.