Playback: use up arrow to dismiss controls and fix for frame rate
Up arrow can now be used to dismiss the OSD controls. This avoids the
problem where you can hit the back button after it has already started
to close them and you end up stopping the playback.

Frame rate for setting refresh rate will be set at 29.97 if there
is a mixture of 23.976 and 29.97 frame rates.