Fix for app crash deu to connection being closed
Occasionally the app would crash with this error:
java.lang.IllegalStateException: Cannot perform this operation
because the connection pool has been closed.

This happens because asynchronous tasks are clsong the database,
sometimes in the middle of another task using it.

Solution is remove all database closes and close it when the app
terminates.