When playing a recording in progress, always set bookmark on exit
At start of playback check if file size is increasing. If so, always set
a bookmark on exit or pause, even if you are at the end. This does not
affect the watched indicator which will be set upon exit if you are
near the end as interpreted by the program.