Support Last Play Position and Bookmarks.
V32 of mythfrontend supports playback position saving and resume from
Last play position as well as supporting a bookmark.

This adds the same support to leanfront. To use it you need a backend
that has the new APIs for Last Play Position, and in V32 you need to
change your port number in leanfront to 6744 instead of 6544. At some
time in the future the port will revert to 6544. This will likely be in V33.