Handle destroyed activities correctly
Activities and fragments can be destroyed when a new activity is moved to
the front, for example the settings page. The code had assumed that this
does not happen. Fixing this required changing a bunch of code, where
application data was stored in a fragment which can be destroyed. It is
now changed to use application context instead of the stored main
activity, which can be destroyed.

To test this set "Don't keep activities" in developer options in Android.

Fixes #34