Handle Error in playback with Detroit PBS station
When playing some recordings from Detroit PBS OTA, an ArrayIndexOutOfBoundsException
occurs in Ac3Util.parseAc3SyncframeInfo.

It is not clear why this happens, the file plays correctly with vlc, mythfrontend
and the exoplayer sample app.

This change modifies exoplayer code to ignore the error and continue playback.
Normally two of these errors occur at the start of playback. If more than
10 of these happen in one playback, abort the playback.