Fix problem with recently viewed list
If you have selected to trim the list so only the latest viewed episode
of a series is included, and you ask the latest viewed episode to be
removed from the list, it was showing the previously viewed episode,
thus making it very difficult to get a show off the list. Fixed so
that if the latest viewed show is removed from the list it will not
include the prior viewed show.