New Feature: "Merge Videos with Recordings
There is a new setting "Merge Videos with Recordings". When enabled,
Videos from the Videos storage group in MythTV will be merged into their
corresponding series in the recording group display. This only applies
if you have recordings of a series and an episode or episodes are in the
videos section, for example if you missed a recording then got it
somewhere else and added it to Videos. You have to make sure you have a
title in the Video metadata that matches the series title. It will
be displayed in the correct place in the episode list based on the
series and episode numbers.

Videos will also be included in the "Related videos" that show in the
details page for a recording. This is regardless of whether "Merge Videos
with Recordings" is enabled.