Support MythTV Commercial Skip
If a recording has been scanned with MythTV commercial detect, leanfront
can use this data to skip commercials or to notify of the presence of
commercials, with the option to skip any commercial found. These features
are enabled in settings and can also be enabled or disabled during playback
using the playback menu.