Fix crash when connecting a keyboard
On Android TV models but not on Fire TV, connecting a keyboard caused a
crash if you were in settings, schedule edit or playback.

This is caused by a configuration change and Android then re-creates
the activity. The application need to be able to handle this.

In settings this is now handled correctly, the activity is restarted
successfully using a saved instance state.

In schedule edit and playback, the recreation of activity is suppressed,
otherwise the saving of the instance state would be excessively complex
and rather unnecessary.