Fix crash on large database of 60,000 recordings
The application will now handle a system with 60000 or more recordings
or videos. However it is extremely slow. Added a setting to limit
download to a lower nulber, defaulted at 10000. It will always load the
latest recordings and leave out older ones if there are too many.
The user can change that setting as needed.