Fix intermittent crash
The program would intermittently crash when hitting the end of a
playback or after performing an update on the video details screen.

java.lang.IllegalStateException: Can not perform this action after
onSaveInstanceState

This was caused by asyncgronour processes invoking a refresh on the main
screen when the main screen had been discarded by teh system.

Fixed by clearing mActiveFragment, which is used to access the main
screen from asynchronous processes.