Live TV rewrite part
LIve TV can be invoked from the program guide or from the Live TV group.
In either case it will now create a recording with the name and
description of the current program on that channel. If there is no
current program in the guide it will assume a program that starts and
ends on the half-hour.

It will stop playback at the end of the program.